refactor: extract previewSavedSearch files to dedicated dir
